moving related classes into email package
