+++ src/main/java/seedu/duke/email/EmailList.java
@@ -1,4 +1,7 @@
-package seedu.duke;
+package seedu.duke.email;
+
+import seedu.duke.Parser;
+import seedu.duke.email.Email;
 
 import java.io.IOException;
 import java.util.ArrayList;
@@ -12,7 +15,7 @@ public class EmailList extends ArrayList<Email> {
         }
         int index = 0;
         String listOfEmails = "This is your list of emails " + "(total of " + this.size() + "): ";
-        for(Email email : this) {
+        for (Email email : this) {
             listOfEmails += "\n" + (++index) + ". " + email.getTitle();
         }
         return listOfEmails;
