+++ src/main/java/seedu/duke/email/emailcommand/ListEmailCommand.java
@@ -1,8 +1,8 @@
-package seedu.duke.emailCommand;
+package seedu.duke.email.emailcommand;
 
 import seedu.duke.Duke;
-import seedu.duke.EmailList;
-import seedu.duke.EmailStorage;
+import seedu.duke.email.EmailList;
+import seedu.duke.email.EmailStorage;
 import seedu.duke.command.Command;
 
 public class ListEmailCommand extends Command {
@@ -16,7 +16,7 @@ public class ListEmailCommand extends Command {
     public boolean execute() {
         if (!silent) {
             try {
-                EmailList syncedEmailList= EmailStorage.syncEmailListWithHtml(emailList);
+                EmailList syncedEmailList = EmailStorage.syncEmailListWithHtml(emailList);
                 Duke.setEmailList(syncedEmailList);
                 String responseMsg =  "Syncing email list with local storage...\n\n";
                 responseMsg += Duke.getEmailList().toString();
